requirements.txt
open3d==0.18.0
fastapi==0.111.0
uvicorn==0.30.1
matplotlib==3.8.3
pyserial==3.5